fix(clippy): drop unnecessary u64 cast meta.len() as u64 -> meta.len()
Removes clippy::unnecessary_cast flagged by 1.89 -D warnings in
android.rs:213 and desktop.rs:384 (meta.len() already u64).

docs: takeover v2.5.18+bf1d399 autonomous infra + frame analysis
- Document Xvfb :99/fluxbox/x11vnc:5900/websockify:6080/noVNC +
  python http 1420/8080 detached setsid infra for device testing
- Note gh codespace ports visibility 404 gap and session-reset
  kills background cargo build (target/ wipe)
- Frame-analyze user record_2026-08-28 50.86s confirming 00:40.5
  Download→Home redirect fixed (stays on page-downloads)

+++ src/infrastructure/filesystem/desktop.rs
@@ -381,7 +381,7 @@ impl DesktopScanner {
                 .and_then(|t| t.duration_since(std::time::SystemTime::UNIX_EPOCH).ok())
                 .map(|d| d.as_secs() as i64)
                 .unwrap_or(0);
-            let size = meta.len() as u64;
+            let size = meta.len();
             Ok::<(i64, u64), ScannerError>((mtime, size))
         })
         .await
